requirements.txt
codeclimate-test-reporter==0.2.1
coverage==4.3.4
coveralls==1.1
dj-database-url==0.4.2
Django==1.11
django-braces==1.11.0
django-dotenv==1.4.1
django-getenv==1.3.2
djangorestframework==3.6.2
docopt==0.6.2
gunicorn==19.7.1
psycopg2==2.7.1
requests==2.13.0
whitenoise==3.3.0
django-oauth-toolkit==0.12.0
django-cors-middleware==1.3.1
